Restructure components into folder-based naming pattern
