+++ src/NextButton/NextButton.css.ts
@@ -1,0 +1,5 @@
+import { style } from '@vanilla-extract/css';
+
+export const container = style({
+  padding: 10
+});
