+++ src/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import App from "./App";
+import App from "./App/App";
 import React from "react";
 
 /** React Version 18 has included updates where ReactDOM
